+++ crates/state/src/settings.rs
@@ -179,6 +179,7 @@ struct Values {
     version: u32,
     normalisation: bool,
     gapless: bool,
+    sleep_timer: bool,
     lyrics_for_local_files: bool,
     karaoke_lyrics: bool,
     blur_lyrics: bool,
@@ -229,6 +230,7 @@ impl Default for Values {
             version: SETTINGS_VERSION,
             normalisation: false,
             gapless: true,
+            sleep_timer: true,
             lyrics_for_local_files: true,
             karaoke_lyrics: true,
             blur_lyrics: true,
@@ -526,6 +528,10 @@ impl AppSettings {
         self.values.gapless
     }
 
+    pub fn sleep_timer(&self) -> bool {
+        self.values.sleep_timer
+    }
+
     pub fn lyrics_for_local_files(&self) -> bool {
         self.values.lyrics_for_local_files
     }
@@ -740,6 +746,11 @@ impl AppSettings {
 
     pub fn set_gapless(&mut self, gapless: bool, cx: &mut Context<Self>) {
         self.values.gapless = gapless;
+        self.schedule_save(cx);
+    }
+
+    pub fn set_sleep_timer(&mut self, sleep_timer: bool, cx: &mut Context<Self>) {
+        self.values.sleep_timer = sleep_timer;
         self.schedule_save(cx);
     }